refactor(recipe-list): Rename list to grid
